Moved BubbleTimer.stories.tsx to /animations
