+++ src/components/animations/BubbleTimer.tsx
