+++ src/components/animations/Loader.tsx
@@ -1,6 +1,5 @@
 import { css } from '@emotion/react';
 import React from 'react';
-
 import AnimatedLoader from '../svg/AnimatedLoader';
 
 export type Props = {}
