+++ src/stories/nrn/animations/BubbleTimer.stories.tsx
@@ -3,7 +3,7 @@ import {
   Story,
 } from '@storybook/react/types-6-0';
 import React from 'react';
-import BubbleTimer, { Props } from '../../../components/utils/BubbleTimer';
+import BubbleTimer, { Props } from '../../../components/animations/BubbleTimer';
 
 export default {
   title: 'Next Right Now/Animations/BubbleTimer',
